+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/servicetemplates/topologytemplates/TopologyTemplateResource.java
@@ -44,6 +44,8 @@ import org.eclipse.winery.repository.rest.RestUtils;
 import org.eclipse.winery.repository.rest.resources._support.dataadapter.composeadapter.CompositionData;
 import org.eclipse.winery.repository.rest.resources.servicetemplates.ServiceTemplateResource;
 import org.eclipse.winery.repository.splitting.Splitting;
+import org.eclipse.winery.repository.targetallocation.Allocation;
+import org.eclipse.winery.repository.targetallocation.util.AllocationRequest;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -216,6 +218,25 @@ public class TopologyTemplateResource {
         }
         URI url = uriInfo.getBaseUri().resolve(RestUtils.getAbsoluteURL(matchedServiceTemplateId));
         return Response.created(url).build();
+    }
+
+    @Path("allocate")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    @POST
+    public Response allocate(@Context UriInfo uriInfo, AllocationRequest allocationRequest) {
+        try {
+            Allocation allocation = new Allocation(allocationRequest);
+            List<ServiceTemplateId> allocatedIds = allocation.allocate((ServiceTemplateId) this.serviceTemplateRes.getId());
+            List<URI> urls = new ArrayList<>();
+            for (ServiceTemplateId id : allocatedIds) {
+                urls.add(uriInfo.getBaseUri().resolve(RestUtils.getAbsoluteURL(id)));
+            }
+            return Response.ok(urls, MediaType.APPLICATION_JSON).build();
+        } catch (Exception e) {
+            LOGGER.debug("Error allocating", e);
+            return Response.serverError().entity(e.getMessage()).build();
+        }
     }
 
     @POST